+++ src/test_geometry_fixtures.rs
@@ -1,134 +1,6 @@
 // FIXME: move to data fixture files once we are able to serialize and deserialize stuff.
 
-use nalgebra::{Point3, Vector3};
-
-use crate::geometry::{Geometry, TriangleFace};
-
-#[allow(dead_code)]
-fn v(x: f32, y: f32, z: f32, translation: [f32; 3], scale: f32) -> Point3<f32> {
-    Point3::new(
-        scale * x + translation[0],
-        scale * y + translation[1],
-        scale * z + translation[2],
-    )
-}
-
-#[allow(dead_code)]
-fn n(x: f32, y: f32, z: f32) -> Vector3<f32> {
-    Vector3::new(x, y, z)
-}
-
-#[allow(dead_code)]
-pub fn quad() -> (Vec<(u32, u32, u32)>, Vec<Point3<f32>>) {
-    #[rustfmt::skip]
-        let vertices = vec![
-        v(-1.0, -1.0, 0.0, [0.0, 0.0, 0.0], 1.0),
-        v(1.0, -1.0, 0.0, [0.0, 0.0, 0.0], 1.0),
-        v(1.0, 1.0, 0.0, [0.0, 0.0, 0.0], 1.0),
-        v(-1.0, 1.0, 0.0, [0.0, 0.0, 0.0], 1.0),
-    ];
-
-    #[rustfmt::skip]
-        let faces = vec![
-        (0, 1, 2),
-        (2, 3, 0),
-    ];
-
-    (faces, vertices)
-}
-
-#[allow(dead_code)]
-pub fn non_manifold_shape() -> (Vec<(u32, u32, u32)>, Vec<Point3<f32>>) {
-    #[rustfmt::skip]
-        let vertices = vec![
-        v(-1.0, -1.0, 0.0, [0.0, 0.0, 0.0], 1.0),
-        v(1.0, -1.0, 0.0, [0.0, 0.0, 0.0], 1.0),
-        v(1.0, 1.0, 0.0, [0.0, 0.0, 0.0], 1.0),
-        v(-1.0, 1.0, 0.0, [0.0, 0.0, 0.0], 1.0),
-        v(0.0, 0.0, 1.0, [0.0, 0.0, 0.0], 1.0),
-    ];
-
-    #[rustfmt::skip]
-        let faces = vec![
-        (0, 1, 2),
-        (2, 3, 0),
-        (2, 4, 0),
-    ];
-
-    (faces, vertices)
-}
-
-#[allow(dead_code)]
-pub fn cube_sharp_mismatching_winding(position: [f32; 3], scale: f32) -> Geometry {
-    let vertex_positions = vec![
-        // back
-        v(-1.0, 1.0, -1.0, position, scale),
-        v(-1.0, 1.0, 1.0, position, scale),
-        v(1.0, 1.0, 1.0, position, scale),
-        v(1.0, 1.0, -1.0, position, scale),
-        // front
-        v(-1.0, -1.0, -1.0, position, scale),
-        v(-1.0, -1.0, 1.0, position, scale),
-        v(1.0, -1.0, 1.0, position, scale),
-        v(1.0, -1.0, -1.0, position, scale),
-    ];
-
-    let vertex_normals = vec![
-        // back
-        n(0.0, 1.0, 0.0),
-        n(0.0, 1.0, 0.0),
-        n(0.0, 1.0, 0.0),
-        n(0.0, 1.0, 0.0),
-        // front
-        n(0.0, -1.0, 0.0),
-        n(0.0, -1.0, 0.0),
-        n(0.0, -1.0, 0.0),
-        n(0.0, -1.0, 0.0),
-        // top
-        n(0.0, 0.0, 1.0),
-        n(0.0, 0.0, 1.0),
-        n(0.0, 0.0, 1.0),
-        n(0.0, 0.0, 1.0),
-        // bottom
-        n(0.0, 0.0, -1.0),
-        n(0.0, 0.0, -1.0),
-        n(0.0, 0.0, -1.0),
-        n(0.0, 0.0, -1.0),
-        // right
-        n(1.0, 0.0, 0.0),
-        n(1.0, 0.0, 0.0),
-        n(1.0, 0.0, 0.0),
-        n(1.0, 0.0, 0.0),
-        // left
-        n(-1.0, 0.0, 0.0),
-        n(-1.0, 0.0, 0.0),
-        n(-1.0, 0.0, 0.0),
-        n(-1.0, 0.0, 0.0),
-    ];
-
-    let faces = vec![
-        // back
-        TriangleFace::new(2, 1, 0),
-        TriangleFace::new(2, 3, 0),
-        // top
-        TriangleFace::new(2, 1, 5),
-        TriangleFace::new(2, 5, 6),
-        // right
-        TriangleFace::new(2, 6, 7),
-        TriangleFace::new(7, 3, 2),
-        // bottom
-        TriangleFace::new(3, 7, 4),
-        TriangleFace::new(4, 0, 3),
-        // front
-        TriangleFace::new(6, 4, 7),
-        TriangleFace::new(4, 6, 5),
-        // left
-        TriangleFace::new(0, 4, 5),
-        TriangleFace::new(5, 1, 0),
-    ];
-
-    Geometry::from_triangle_faces_with_vertices_and_normals(faces, vertex_positions, vertex_normals)
-}
+use nalgebra::Point3;
 
 #[allow(dead_code)]
 pub fn torus() -> (Vec<(u32, u32, u32)>, Vec<Point3<f32>>) {
